requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.3
python-dotenv>=1.0.1
lxml>=5.0.0
pytest>=8.1.1
pytest-mock>=3.14.0
responses>=0.25.3
